+++ src/components/FeedbackForm/feedback-form.css
@@ -1,0 +1,7 @@
+.overview form {
+    width: 100%;
+}
+
+.overview form textarea {
+    resize: none;
+}
