+++ src/components/FeedbackForm/index.js
@@ -1,0 +1,3 @@
+import FeedbackForm from './FeedbackForm';
+
+export default FeedbackForm;
