+++ src/components/TaskOverview/task-overview.css
@@ -2,14 +2,6 @@
     font-weight: 600;
 }
 
-.overview textarea {
-    resize: none;
-}
-
-.overview form {
-    width: 100%;
-}
-
 .overview button {
     max-width: 122px;
     position: relative;
